Add the removal of a single feature to the demo
Now a feature can be removed from the results set from
a select. There was a state issue that required adding a
revision counter to the reducer.

+++ src/gm3/application.js
@@ -576,10 +576,22 @@ class Application {
 
     /** Removes features from a query.
      */
-    removeQueryResults(queryId, filter) {
-        this.store.dispatch(mapActions.queryProgress(queryId));
-        this.store.dispatch(mapActions.removeQueryResults(queryId, filter));
-        this.store.dispatch(mapActions.finishQuery(queryId));
+    removeQueryResults(queryId, filter, options = {}) {
+        const execRemove = choice => {
+            if (choice === 'confirm') {
+                this.store.dispatch(mapActions.queryProgress(queryId));
+                this.store.dispatch(mapActions.removeQueryResults(queryId, filter));
+                // also remove the features from the results layer.
+                this.removeFeatures('results/results', filter);
+                this.store.dispatch(mapActions.finishQuery(queryId));
+            }
+        };
+
+        if (options.withConfirm) {
+            this.confirm('remove-feature', 'Remove feature?', execRemove);
+        } else {
+            execRemove('confirm');
+        }
     }
 
     /** Remove features from a vector layer (with a filter)
